ghzawi

+++ fisrt-task/database/migrations/2024_02_09_150509_create_attributes_table.php
@@ -19,9 +19,9 @@ return new class extends Migration
             $table->string('name_en');
             $table->unsignedBigInteger('product_id');
             $table->foreign('product_id')->references('id')->on('products'); 
-            $table->float('price');
+            $table->integer('price');
             $table->string('discount_type');
-            $table->float('discount_amount'); 
+            $table->integer('discount_amount'); 
             $table->integer('quantity');
             $table->timestamps();
         });
